requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.2
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cv==1.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.7
et-xmlfile==1.0.1
greenlet==1.0.0
idna==2.10
image==1.5.33
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonpath==0.82
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.4.1
nacos-sdk-python==0.1.6
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.7
pandas==1.2.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.17
pyee==8.1.0
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pypiwin32==223
pyppeteer==0.2.5
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
requests==2.25.1
retrying==1.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.0
sqlparse==0.4.1
tqdm==4.60.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.4
wcwidth==0.2.5
websockets==8.1
xlwt==1.3.0
yamail==1.0.2
